Long variation ID

+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/cart/request/OrderForCalculate.kt
@@ -111,7 +111,7 @@ data class OrderForCalculate(
 		val giftCard: Boolean? = null,
 
 		val selectedOptions: List<OrderItemOption>? = null,
-		val combinationId: Int? = null,
+		val combinationId: Long? = null,
 		val taxes: List<OrderItemTax>? = null,
 		val files: List<OrderItemProductFile>? = null,
 		val dimensions: ProductDimensions? = null,
