+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/cart/result/CalculateOrderDetailsResult.kt
@@ -166,7 +166,7 @@ data class CalculateOrderDetailsResult(
 		val giftCard: Boolean? = null,
 
 		val selectedOptions: List<OrderItemOption>? = null,
-		val combinationId: Int? = null,
+		val combinationId: Long? = null,
 		val taxes: List<OrderItemTax>? = null,
 		val files: List<OrderItemProductFile>? = null,
 		val dimensions: ProductDimensions? = null,
