+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/product/result/FetchedProduct.kt
@@ -120,7 +120,7 @@ data class FetchedProduct(
 	val files: List<ProductFile>? = null,
 	val favorites: FavoritesStats? = null, // TODO Figure out how to test
 
-	val defaultCombinationId: Int? = null, // TODO implement combinations support
+	val defaultCombinationId: Long? = null, // TODO implement combinations support
 	val combinations: List<FetchedVariation>? = null, // TODO implement combinations support
 	val isGiftCard: Boolean? = null, // TODO from ECWID-67826: this flag is read-only. API for creating gift cards in the plans (BACKLOG-4157)
 	val discountsAllowed: Boolean? = null,
