+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/variation/request/ProductVariationImageAsyncUploadRequest.kt
@@ -7,7 +7,7 @@ import com.ecwid.apiclient.v3.impl.RequestInfo
 
 data class ProductVariationImageAsyncUploadRequest(
 	val productId: Int = 0,
-	val variationId: Int = 0,
+	val variationId: Long = 0,
 	val asyncPictureData: AsyncPictureData = AsyncPictureData()
 ) : ApiRequest {
 	override fun toRequestInfo(): RequestInfo {
