+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/variation/request/ProductVariationImageUploadRequest.kt
@@ -8,7 +8,7 @@ import com.ecwid.apiclient.v3.impl.RequestInfo
 
 data class ProductVariationImageUploadRequest(
 	val productId: Int = 0,
-	val variationId: Int = 0,
+	val variationId: Long = 0,
 	val fileData: UploadFileData = UploadFileData.ExternalUrlData("")
 ) : ApiRequest {
